refactor: extract board verdicts module
Resolves the temporary circular imports from data-readers.mjs and
metrics.mjs (both now import readVerdicts/readPlanCosts/readQAStats/
readSecStats from lib/verdicts.mjs directly instead of ../server.mjs).
No more TEMPORARY export markers or circular imports remain in the tree.
